feat: add Links, route-moderation page, settings page
